fix: Creating edges is blocked in read-only mode (WB-339)
