+++ .changeset/pretty-months-thank.md
@@ -1,0 +1,5 @@
+---
+'@workflowbuilder/sdk': minor
+---
+
+fix: creating edges is blocked in read-only mode.
