+++ packages/sdk/src/features/diagram/selectors.ts
@@ -8,7 +8,6 @@ export function diagramStateSelector({
   edges,
   onNodesChange,
   onEdgesChange,
-  onConnect,
   onInit,
   isReadOnlyMode,
   onEdgeMouseEnter,
@@ -20,7 +19,6 @@ export function diagramStateSelector({
     edges,
     onNodesChange,
     onEdgesChange,
-    onConnect,
     onInit,
     isReadOnlyMode,
     onEdgeMouseEnter,
